chore: update decimal column in migrations

+++ database/migrations/2021_12_13_182904_create_shop_order_items_table.php
@@ -4,7 +4,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class () extends Migration {
+return new class() extends Migration
+{
     /**
      * Run the migrations.
      *
@@ -17,7 +18,7 @@ return new class () extends Migration {
             $table->foreignId('shop_order_id')->nullable()->constrained()->cascadeOnDelete();
             $table->foreignId('shop_product_id')->nullable()->constrained()->cascadeOnDelete();
             $table->integer('qty');
-            $table->integer('unit_price');
+            $table->decimal('unit_price', 10, 2);
             $table->timestamps();
         });
     }
